finished mutation connection

+++ gql-prisma/src/resolvers/Mutation.js
@@ -80,17 +80,34 @@ const Mutation = {
 		return prisma.mutation.deletePost({where: {id: args.id}}, info)
 		
 	},
-	updatePost(parent, args, { prisma }, info) {
+	async updatePost(parent, args, { prisma, request }, info) {
 
-		return prisma.mutation.updatePost({data: {...args.data, author: {connect: args.author}}, where: {id: args.id}}, info)
+        const userId = getUserId(request)
+
+        const postExists = await prisma.exists.Post({
+            id: args.id,
+            author: {
+                id: userId
+            }
+        })
+
+        if (!postExists) {
+            throw new Error('Unable to delete post')
+        }
+
+        return prisma.mutation.updatePost({data: args.data, 
+            where: {id: args.id}}, info)
 	},
-	createComment(parent, args, { prisma }, info) {
+	createComment(parent, args, { prisma, request }, info) {
+
+        const userId = getUserId(request)
+
         return prisma.mutation.createComment({
             data: {
                 text: args.data.text,
                 author: {
                     connect: {
-                        id: args.data.author
+                        id: userId
                     }
                 },
                 post: {
@@ -101,14 +118,38 @@ const Mutation = {
             }
         }, info)
     },
-    deleteComment(parent, args, { prisma }, info) {
+    async deleteComment(parent, args, { prisma, request }, info) {
+
+        const userId = getUserId(request)
+
+        const commentExists = await prisma.exists.Comment({
+            id: args.id,
+            author: {id: userId}
+        })
+
+        if (!commentExists) {
+            throw new Error('Unable to delete comment')
+        }
+
         return prisma.mutation.deleteComment({
             where: {
                 id: args.id
             }
         }, info)
     },
-    updateComment(parent, args, { prisma }, info) {
+    async updateComment(parent, args, { prisma, request }, info) {
+
+        const userId = getUserId(request)
+
+        const commentExists = await prisma.exists.Comment({
+            id: args.id,
+            author: {id: userId}
+        })
+
+        if (!commentExists) {
+            throw new Error('Unable to update comment')
+        }
+
         return prisma.mutation.updateComment({
             where: {
                 id: args.id
